+++ tests/entries/unit.mjs
@@ -836,6 +836,7 @@ for (PATH of ['core-js-pure', 'core-js']) {
     ` === 'a1b');
     ok('next' in load(NS, 'string/code-points')('a'));
     ok('next' in load(NS, 'string/virtual/code-points').call('a'));
+    ok(load(NS, 'symbol/custom-matcher'));
     ok(load(NS, 'symbol/is-registered-symbol')(1) === false);
     ok(load(NS, 'symbol/is-well-known-symbol')(1) === false);
     ok(load(NS, 'symbol/is-registered')(1) === false);
@@ -948,6 +949,7 @@ for (PATH of ['core-js-pure', 'core-js']) {
   load('proposals/object-values-entries');
   load('proposals/observable');
   load('proposals/pattern-matching');
+  load('proposals/pattern-matching-v2');
   load('proposals/promise-all-settled');
   load('proposals/promise-any');
   load('proposals/promise-finally');